requirements.txt
huggingface_hub==0.23.1
importlib-metadata==7.0.1
matplotlib==3.8.4
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
Pillow==10.4.0
ply==3.11
pyparsing==3.0.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.1
setuptools==69.5.1
shap==0.42.1
tokenizers==0.15.1
torchaudio==2.3.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.37.2
